Test: Passed all tests
Fixed bug in which messages were not calculated properly in the test. Fixed bug in which messages from node were not calculated correctly in the code (both cuda and cpu versions). More specifically for the calculation of the nodes messages previously the previous_messages tensor was used, while it should have been the new_messages tensor.